Streamline FindNodes requests
