+++ src/handler/session.rs
@@ -60,7 +60,6 @@ impl Session {
 
         // If the message nonce length is ever set below 4 bytes this will explode. The packet
         // size constants shouldn't be modified.
-        debug_assert!(MESSAGE_NONCE_LENGTH > 4);
         let random_nonce: [u8; MESSAGE_NONCE_LENGTH - 4] = rand::random();
         let mut message_nonce: MessageNonce = [0u8; crate::packet::MESSAGE_NONCE_LENGTH];
         message_nonce[..4].copy_from_slice(&self.counter.to_be_bytes());
@@ -69,7 +68,7 @@ impl Session {
         // the authenticated data is the packet header
         let header = PacketHeader {
             src_id,
-            flag: PacketType::Message(message_nonce.clone()),
+            flag: PacketType::Message(message_nonce),
         };
 
         let cipher = crypto::encrypt_message(
@@ -199,7 +198,7 @@ impl Session {
         // build an authentication packet
         let message_nonce: MessageNonce = rand::random();
         let mut packet = Packet::new_authheader(
-            local_node_id.clone(),
+            *local_node_id,
             message_nonce,
             sig,
             ephem_pubkey,
