+++ src/kbucket.rs
@@ -241,30 +241,40 @@ where
     }
 
     /// Returns an iterator over the keys that are contained in a kbucket, specified by a log2 distance.
-    pub fn nodes_by_distance<'a>(
+    pub fn nodes_by_distances<'a>(
         &'a mut self,
-        log2_distance: u64,
+        log2_distances: Vec<u64>,
     ) -> Vec<EntryRefView<'a, TNodeId, TVal>> {
-        // the log2 distance ranges from 1-256 and is always 1 more than the bucket index. For this
-        // reason we subtract 1 from log2 distance to get the correct bucket index.
-        if log2_distance > 0 && log2_distance <= (NUM_BUCKETS as u64) {
-            let bucket = &mut self.buckets[(log2_distance - 1) as usize];
+        let distances = log2_distances
+            .into_iter()
+            .filter(|&d| d > 0 && d <= (NUM_BUCKETS as u64))
+            .collect::<Vec<_>>();
+
+        // apply bending nodes
+        for distance in &distances {
+            // the log2 distance ranges from 1-256 and is always 1 more than the bucket index. For this
+            // reason we subtract 1 from log2 distance to get the correct bucket index.
+            let bucket = &mut self.buckets[(distance - 1) as usize];
             if let Some(applied) = bucket.apply_pending() {
                 self.applied_pending.push_back(applied)
             }
-            bucket
-                .iter()
-                .map(|(n, status)| {
-                    let node = NodeRefView {
-                        key: &n.key,
-                        value: &n.value,
-                    };
-                    EntryRefView { node, status }
-                })
-                .collect()
-        } else {
-            Vec::new()
         }
+
+        // find the matching nodes
+        let mut matching_nodes = Vec::new();
+        for distance in distances {
+            let bucket = &self.buckets[(distance - 1) as usize];
+            for node in bucket.iter().map(|(n, status)| {
+                let node = NodeRefView {
+                    key: &n.key,
+                    value: &n.value,
+                };
+                EntryRefView { node, status }
+            }) {
+                matching_nodes.push(node);
+            }
+        }
+        matching_nodes
     }
 
     /// Returns an iterator over the keys closest to `target`, ordered by
