+++ src/query_pool.rs
@@ -25,7 +25,7 @@ mod peers;
 
 pub(crate) use peers::closest::{FindNodeQuery, FindNodeQueryConfig};
 pub(crate) use peers::predicate::{PredicateQuery, PredicateQueryConfig};
-pub use peers::{QueryState, ReturnPeer};
+pub use peers::QueryState;
 
 use crate::kbucket::{Key, PredicateKey};
 use fnv::FnvHashMap;
@@ -53,12 +53,7 @@ pub enum QueryPoolState<'a, TTarget, TNodeId, TResult> {
     Idle,
     /// At least one query is waiting for results. `Some(request)` indicates
     /// that a new request is now being waited on.
-    Waiting(
-        Option<(
-            &'a mut Query<TTarget, TNodeId, TResult>,
-            ReturnPeer<TNodeId>,
-        )>,
-    ),
+    Waiting(Option<(&'a mut Query<TTarget, TNodeId, TResult>, TNodeId)>),
     /// A query has finished.
     Finished(Query<TTarget, TNodeId, TResult>),
     /// A query has timed out.
@@ -91,13 +86,12 @@ where
         config: FindNodeQueryConfig,
         target: TTarget,
         peers: I,
-        iterations: usize,
     ) -> QueryId
     where
         I: IntoIterator<Item = Key<TNodeId>>,
     {
         let target_key = target.key();
-        let findnode_query = FindNodeQuery::with_config(config, target_key, peers, iterations);
+        let findnode_query = FindNodeQuery::with_config(config, target_key, peers);
         let peer_iter = QueryPeerIter::FindNode(findnode_query);
         self.add(peer_iter, target)
     }
@@ -108,15 +102,13 @@ where
         config: PredicateQueryConfig,
         target: TTarget,
         peers: I,
-        iterations: usize,
         predicate: impl Fn(&TResult) -> bool + Send + 'static,
     ) -> QueryId
     where
         I: IntoIterator<Item = PredicateKey<TNodeId>>,
     {
         let target_key = target.key();
-        let predicate_query =
-            PredicateQuery::with_config(config, target_key, peers, iterations, predicate);
+        let predicate_query = PredicateQuery::with_config(config, target_key, peers, predicate);
         let peer_iter = QueryPeerIter::Predicate(predicate_query);
         self.add(peer_iter, target)
     }
